requirements.txt
oscpy
PyAudio
PyQt5
matplotlib
scipy
python-osc